pyproject.toml
[project]
name = "gateway"
version = "0.1.0"
description = "Gatweway with Autogen integration for Zep"
# readme = "README2.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn",
    "watchdog==4.0.0",
    "websockets==12.0",
    "pydantic==2.10.4",
    "typing-extensions>=4.5.0",
    "loguru==0.7.2",
    "azure-identity>=1.23.1",
    "aiohttp>=3.12.15",
    "python-dotenv>=1.0.1",
    "python-slugify>=8.0.4",
    "rich>=14.0.0",
    "ag2[openai]",
    "httpx",
    # "ag2[gemini]",
    # "ag2[anthropic,cohere,mistral]",
    # "ag2[ollama]>=0.9",
    # "ollama>=0.4.8",
    "streamlit>=1.44.1",
    "autogen-agentchat>=0.7.1",
    "autogen-ext[azure,openai]>=0.7.1",
    "autogen-core>=0.7.1",
    "zep-cloud>=3.0.5",
    "zep_autogen>=0.1.1",

    "zep-python",
]


[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov",
    "pytest-asyncio",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.uv.pip]
extra-index-url = ["https://pypi.org/simple"]

# [build-system]
# requires = ["hatchling"]
# build-backend = "hatchling.build"

# [tool.hatch.build.targets.wheel]
# packages = ["src/zep_autogen"]

# [tool.pytest.ini_options]
# testpaths = ["tests"]
# python_files = ["test_*.py"]
# python_classes = ["Test*"]
# python_functions = ["test_*"]
# markers = [
#     "integration: marks tests as integration tests (deselect with '-m \"not integration\"')"
# ]

# [tool.ruff]
# target-version = "py310"
# line-length = 100

# [tool.ruff.lint]
# select = [
#     "E",  # pycodestyle errors
#     "W",  # pycodestyle warnings
#     "F",  # pyflakes
#     "I",  # isort
#     "B",  # flake8-bugbear
#     "C4", # flake8-comprehensions
#     "UP", # pyupgrade
# ]
# ignore = [
#     "E501", # line too long, handled by formatter
#     "B008", # do not perform function calls in argument defaults
#     "C901", # too complex
# ]

# [tool.ruff.lint.isort]
# known-first-party = ["zep_autogen"]

# [tool.ruff.format]
# quote-style = "double"
# indent-style = "space"
# skip-magic-trailing-comma = false
# line-ending = "auto"

# [tool.mypy]
# python_version = "3.10"
# warn_return_any = true
# warn_unused_configs = true
# disallow_untyped_defs = true
# disallow_incomplete_defs = true
# check_untyped_defs = true

# [[tool.mypy.overrides]]
# module = [
#     "autogen_core.*",
#     "autogen_agentchat.*", 
#     "autogen_ext.*",
#     "zep_cloud.*",
# ]
# ignore_missing_imports = true
